Extract abstract base class `BaseInterfaceProxyGenerator`
